update client admin

+++ src/main/java/travel/manager/model/home/CategoryTours.java
@@ -11,7 +11,7 @@ public class CategoryTours {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "category_tour_id")
-    private int category_tour_id;
+    private Long category_tour_id;
     private String category_tour_name;
     private String category_tour_url;
